+++ lib/src/configuration.dart
@@ -1,4 +1,5 @@
 import 'dart:io';
+import 'package:path/path.dart' as p;
 import 'package:args/args.dart';
 import 'package:cli_util/cli_logging.dart';
 import 'package:get_it/get_it.dart';
@@ -22,7 +23,7 @@ class Configuration {
   String? msixVersion;
   String? appDescription;
   String buildFilesFolder =
-      '${Directory.current.path}/build/windows/runner/Release';
+      p.join(Directory.current.path, 'build', 'windows', 'runner', 'Release');
   String? certificatePath;
   String? certificatePassword;
   String? publisher;
@@ -59,14 +60,16 @@ class Configuration {
   StartupTask? startupTask;
   Iterable<String>? appUriHandlerHosts;
   Iterable<String>? languages;
-  String get defaultsIconsFolderPath => '$msixAssetsPath/icons';
-  String get msixToolkitPath => '$msixAssetsPath/MSIX-Toolkit';
+  String get defaultsIconsFolderPath => p.join(msixAssetsPath, 'icons');
+  String get msixToolkitPath =>
+      p.join(msixAssetsPath, 'MSIX-Toolkit', 'Redist.x64');
   String get msixPath =>
-      '${outputPath ?? buildFilesFolder}/${outputName ?? appName}.msix';
-  String get appInstallerPath =>
-      '$publishFolderPath/${basename(msixPath).replaceAll('.msix', '.appinstaller')}';
+      p.join(outputPath ?? buildFilesFolder, '${outputName ?? appName}.msix');
+  String get appInstallerPath => p.join(publishFolderPath!,
+      basename(msixPath).replaceAll('.msix', '.appinstaller'));
   String pubspecYamlPath = "pubspec.yaml";
   String osMinVersion = '10.0.17763.0';
+  bool isTestCertificate = false;
 
   Configuration(this._arguments);
 
@@ -106,10 +109,8 @@ class Configuration {
         yaml['store']?.toString().toLowerCase() == 'true';
     createWithDebugBuildFiles = _args.wasParsed('debug') ||
         yaml['debug']?.toString().toLowerCase() == 'true';
-    if (createWithDebugBuildFiles) {
-      buildFilesFolder = buildFilesFolder.replaceFirst('Release', 'Debug');
-    }
 
+    displayName = _args['display-name'] ?? yaml['display_name'];
     displayNames = (yaml['app_names'] as YamlMap?)?.value;
     descriptions = (yaml['app_descriptions'] as YamlMap?)?.value;
 
@@ -234,6 +235,14 @@ class Configuration {
     }
     if (msixVersion.isNull) msixVersion = '1.0.0.0';
     if (architecture.isNull) architecture = 'x64';
+
+    buildFilesFolder = await _getBuildFilesFolder();
+
+    if (createWithDebugBuildFiles) {
+      buildFilesFolder = buildFilesFolder.replaceFirst('Release', 'Debug');
+    }
+    _logger.trace('pack files in build folder: $buildFilesFolder');
+
     languages ??= ['en-us'];
 
     if (!RegExp(r'^(\*|\d+(\.\d+){3,3}(\.\*)?)$').hasMatch(msixVersion!)) {
@@ -263,12 +272,13 @@ class Configuration {
       }
     } else {
       // if no certificate was chosen then use test certificate
-      certificatePath = '$msixAssetsPath/test_certificate.pfx';
+      certificatePath = p.join(msixAssetsPath, 'test_certificate.pfx');
       certificatePassword = '1234';
+      isTestCertificate = true;
     }
 
-    if (!['x86', 'x64'].contains(architecture)) {
-      throw 'Architecture can be "x86" or "x64", check "msix_config: architecture" at pubspec.yaml';
+    if (!['x64', 'arm64'].contains(architecture)) {
+      throw 'Architecture can be "x64" or "arm64", check "msix_config: architecture" at pubspec.yaml';
     }
   }
 
@@ -339,6 +349,29 @@ class Configuration {
     _args = parser.parse(args.where((arg) => arg != '-v'));
   }
 
+  Future<String> _getBuildFilesFolder() async {
+    final String buildFilesFolderStart =
+        buildFilesFolder.substring(0, buildFilesFolder.indexOf('runner'));
+    final String buildFilesFolderArchitecture =
+        p.join(buildFilesFolderStart, architecture);
+    final String buildFilesFolderEnd =
+        buildFilesFolder.substring(buildFilesFolder.indexOf('runner'));
+
+    if (architecture == 'arm64' &&
+        !(await Directory(buildFilesFolderArchitecture).exists())) {
+      _logger.stderr(
+          'cannot find arm64 build folder at: $buildFilesFolderArchitecture');
+      exit(-1);
+    }
+
+    if (await Directory(p.join(buildFilesFolderArchitecture, 'runner'))
+        .exists()) {
+      return p.join(buildFilesFolderStart, architecture, buildFilesFolderEnd);
+    } else {
+      return p.join(buildFilesFolderStart, buildFilesFolderEnd);
+    }
+  }
+
   Future<void> validateAppInstallerConfigValues() async {
     _logger.trace('validating app installer config values');
 
@@ -363,8 +396,19 @@ class Configuration {
       throw 'Failed to locate or read package config.';
     }
 
-    Package msixPackage = packagesConfig.packages
-        .firstWhere((package) => package.name == "mg_msix");
+    Package? msixPackage = packagesConfig['mg_msix'];
+
+    // Locate package config from script file directory
+    if (msixPackage == null) {
+      final scriptFile = File.fromUri(Platform.script);
+      packagesConfig = await findPackageConfig(scriptFile.parent);
+      msixPackage = packagesConfig?['msix'];
+    }
+
+    if (msixPackage == null) {
+      throw 'Failed to locate msix assets path.';
+    }
+
     String path =
         '${msixPackage.packageUriRoot.toString().replaceAll('file:///', '')}assets';