+++ test/appx_manifest_test.dart
@@ -1,12 +1,12 @@
 import 'dart:io';
-
+import 'package:path/path.dart' as p;
 import 'package:cli_util/cli_logging.dart';
 import 'package:get_it/get_it.dart';
 import 'package:mg_msix/src/appx_manifest.dart';
 import 'package:mg_msix/src/configuration.dart';
 import 'package:test/test.dart';
 
-const tempFolderPath = 'test/appx_manifest_temp';
+var tempFolderPath = p.join('test', 'runner', 'appx_manifest_temp');
 
 void main() {
   late Configuration config;
@@ -39,20 +39,21 @@ void main() {
 
     GetIt.I.registerSingleton<Configuration>(config);
 
-    await Directory('$tempFolderPath/').create(recursive: true);
+    await Directory(tempFolderPath).create(recursive: true);
   });
 
   tearDown(() async {
     GetIt.I.reset();
 
-    if (await Directory('$tempFolderPath/').exists()) {
-      await Directory('$tempFolderPath/').delete(recursive: true);
+    if (await Directory(tempFolderPath).exists()) {
+      await Directory(tempFolderPath).delete(recursive: true);
     }
   });
 
   test('manifest created', () async {
     await AppxManifest().generateAppxManifest();
-    expect(await File('$tempFolderPath/AppxManifest.xml').exists(), true);
+    expect(
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).exists(), true);
   });
 
   test('identityName is valid', () async {
@@ -60,7 +61,7 @@ void main() {
     config.identityName = testValue;
     await AppxManifest().generateAppxManifest();
     expect(
-        (await File('$tempFolderPath/AppxManifest.xml').readAsString())
+        (await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString())
             .contains('<Identity Name="$testValue"'),
         true);
   });
@@ -70,7 +71,7 @@ void main() {
     config.publisher = testValue;
     await AppxManifest().generateAppxManifest();
     expect(
-        (await File('$tempFolderPath/AppxManifest.xml').readAsString())
+        (await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString())
             .contains('Publisher="$testValue"'),
         true);
   });
@@ -80,7 +81,7 @@ void main() {
     config.publisherName = testValue;
     await AppxManifest().generateAppxManifest();
     expect(
-        (await File('$tempFolderPath/AppxManifest.xml').readAsString())
+        (await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString())
             .contains(
                 '<PublisherDisplayName>$testValue</PublisherDisplayName>'),
         true);
@@ -91,7 +92,7 @@ void main() {
     config.msixVersion = testValue;
     await AppxManifest().generateAppxManifest();
     expect(
-        (await File('$tempFolderPath/AppxManifest.xml').readAsString())
+        (await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString())
             .contains('Version="$testValue"'),
         true);
   });
@@ -101,7 +102,7 @@ void main() {
     config.appName = testValue;
     await AppxManifest().generateAppxManifest();
     expect(
-        (await File('$tempFolderPath/AppxManifest.xml').readAsString())
+        (await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString())
             .contains('Id="${testValue.replaceAll('_', '')}"'),
         true);
   });
@@ -111,7 +112,7 @@ void main() {
     config.appDescription = testValue;
     await AppxManifest().generateAppxManifest();
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(manifestContent.contains('<Description>$testValue</Description>'),
         true);
     expect(manifestContent.contains('Description="$testValue"'), true);
@@ -122,7 +123,7 @@ void main() {
     config.displayName = testValue;
     await AppxManifest().generateAppxManifest();
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(manifestContent.contains('<DisplayName>$testValue</DisplayName>'),
         true);
     expect(manifestContent.contains('DisplayName="$testValue"'), true);
@@ -135,7 +136,7 @@ void main() {
     config.architecture = testValue;
     await AppxManifest().generateAppxManifest();
     expect(
-        (await File('$tempFolderPath/AppxManifest.xml').readAsString())
+        (await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString())
             .contains('ProcessorArchitecture="$testValue"'),
         true);
   });
@@ -145,7 +146,7 @@ void main() {
     config.executableFileName = testValue;
     await AppxManifest().generateAppxManifest();
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(manifestContent.contains('Executable="$testValue'), true);
   });
 
@@ -155,7 +156,7 @@ void main() {
     await AppxManifest().generateAppxManifest();
 
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(
         manifestContent.contains(
             '<desktop:ExecutionAlias Alias="${testValue.toLowerCase()}.exe" />'),
@@ -167,7 +168,7 @@ void main() {
     config.protocolActivation = [testValue];
     await AppxManifest().generateAppxManifest();
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(manifestContent.contains('<uap:Protocol Name="$testValue">'), true);
     expect(
         manifestContent.contains(
@@ -181,7 +182,7 @@ void main() {
     config.fileExtension = testValue;
     await AppxManifest().generateAppxManifest();
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(
         manifestContent
             .contains('<uap:FileType>.fileExtension_test1</uap:FileType>'),
@@ -204,7 +205,7 @@ void main() {
     await AppxManifest().generateAppxManifest();
 
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(manifestContent.contains('Executable="$testValue'), true);
     expect(
         manifestContent.contains(
@@ -216,7 +217,7 @@ void main() {
     config.capabilities = testValue;
     await AppxManifest().generateAppxManifest();
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(manifestContent.contains('<uap:Capability Name="videosLibrary" />'),
         true);
     expect(manifestContent.contains('<DeviceCapability Name="microphone" />'),
@@ -230,7 +231,7 @@ void main() {
     config.languages = ['en-us', 'he-il'];
     await AppxManifest().generateAppxManifest();
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(manifestContent.contains('<Resource Language="en-us" />'), true);
     expect(manifestContent.contains('<Resource Language="he-il" />'), true);
   });
@@ -240,7 +241,7 @@ void main() {
     config.toastActivatorCLSID = testValue;
     await AppxManifest().generateAppxManifest();
     var manifestContent =
-        await File('$tempFolderPath/AppxManifest.xml').readAsString();
+        await File(p.join(tempFolderPath, 'AppxManifest.xml')).readAsString();
     expect(
         manifestContent.contains(
             '<com:ExeServer Executable="executableFileName_test" Arguments="----AppNotificationActivationServer" DisplayName="Toast activator">'),